Redesign the trace merging behaviour.
When it comes to compiling traces, it should be possible to know how to
convert `TEST` instructions into guards. When collecting traces, the
jit compiler also collects information about which branch the
execution takes. The trace tree now encodes the execution path as well
(right subtree is the "false" branch).